color handlers, fluid type

+++ src/main/java/com/accbdd/aqua_vitae/component/AlcoholPropertiesComponent.java
@@ -21,14 +21,14 @@ import java.util.Set;
  * @param flavors
  * @param items TODO: support fluidstacks
  */
-public record AlcoholPropertiesComponent(int color, int abb, int age, Set<ResourceKey<BrewingIngredient.Flavor>> flavors, List<ItemStack> items) {
+public record AlcoholPropertiesComponent(int color, float abb, int age, Set<ResourceKey<BrewingIngredient.Flavor>> flavors, List<ItemStack> items) {
 
     public static final AlcoholPropertiesComponent EMPTY = new AlcoholPropertiesComponent(0, 0, 0, Set.of(), List.of());
 
     public static final Codec<AlcoholPropertiesComponent> CODEC = RecordCodecBuilder.create(instance ->
             instance.group(
                     Codec.INT.fieldOf("color").forGetter(AlcoholPropertiesComponent::color),
-                    Codec.INT.fieldOf("abb").forGetter(AlcoholPropertiesComponent::abb),
+                    Codec.FLOAT.fieldOf("abb").forGetter(AlcoholPropertiesComponent::abb),
                     Codec.INT.fieldOf("age").forGetter(AlcoholPropertiesComponent::age),
                     ResourceKey.codec(AquaVitae.FLAVOR_REGISTRY).listOf().xmap(Set::copyOf, List::copyOf).fieldOf("flavors").forGetter(AlcoholPropertiesComponent::flavors),
                     ItemStack.CODEC.listOf().fieldOf("items").forGetter(AlcoholPropertiesComponent::items)
@@ -37,7 +37,7 @@ public record AlcoholPropertiesComponent(int color, int abb, int age, Set<Resour
 
     public static final StreamCodec<RegistryFriendlyByteBuf, AlcoholPropertiesComponent> STREAM_CODEC = StreamCodec.composite(
             ByteBufCodecs.INT, AlcoholPropertiesComponent::color,
-            ByteBufCodecs.INT, AlcoholPropertiesComponent::abb,
+            ByteBufCodecs.FLOAT, AlcoholPropertiesComponent::abb,
             ByteBufCodecs.INT, AlcoholPropertiesComponent::age,
             ResourceKey.streamCodec(AquaVitae.FLAVOR_REGISTRY).apply(ByteBufCodecs.collection(NonNullList::createWithCapacity)).map(Set::copyOf, NonNullList::copyOf), AlcoholPropertiesComponent::flavors,
             ItemStack.LIST_STREAM_CODEC, AlcoholPropertiesComponent::items,
